requirements.txt
requests>=2.28.0
pyinaturalist>=0.18.0
Pillow>=9.0.0
pathlib2>=2.3.0 